line:58 修改註解

+++ musical-notation-editor/src/CheckBeat.java
@@ -55,7 +55,7 @@ public class CheckBeat
 				System.out.println("節拍數:"+Beat+" "+"小節數:"+sectionId+" 標準節拍數:"+sectionBeat);//印出是音檔中哪一小節和算出的節拍數
 				/*
 				 可以在這邊check每一音檔每一小節是否有節拍錯誤 
-				if(sectionBeat == Beat)  
+				if(sectionBeat == (int)Beat)  
 				{
 					正確!!!
 				}
